requirements.docker.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
itsdangerous==2.2.0

torch==2.5.1+cpu
transformers==4.44.0
Pillow==10.4.0
PyYAML==6.0.2
accelerate==1.12.0
psutil==7.2.2
